updateUser and deleteUser now connect store

+++ src/components/profile-view/profile-view.jsx
@@ -2,7 +2,7 @@ import React, { useState, useEffect } from 'react';
 import axios from 'axios';
 import { connect } from 'react-redux';
 
-import { setUser, setFavorite } from '../../actions/actions';
+import { updateUser, deleteUser } from '../../actions/actions';
 
 import { Container, Row, CardGroup, Card, Button } from 'react-bootstrap';
 
@@ -12,28 +12,22 @@ import { UpdateUser } from './update-user';
 import './profile-view.scss';
 
 function ProfileView(props) {
-    let { movies, user, unfavorite } = props;
+    let { movies, user } = props;
 
-    const token = localStorage.getItem("token");
+    const token = localStorage.getItem('token');
 
-    const handleSubmit = async (e, username, newPassword, email, user) => {
-        e.preventDefault();
+    const handleUpdate = async (updateObject) => {
         try {
-            let updateObject = {
-                Username: username ? username : user.Username,
-                Email: email ? email : user.Email,
-            };
-            if (newPassword) {
-                updateObject.Password = newPassword
-            }
-            const response = await axios.put(`https://trackm-app.herokuapp.com/users/${user.Username}`, updateObject, {
+            const response = await axios.put(`https://trackm-app.herokuapp.com/users/${user.Username}`, { ...updateObject }, {
                 headers: { Authorization: `Bearer ${token}` },
             });
+            console.log('reponse', response);
+            updateUser({ ...updateObject });
             console.log(updateObject);
-            if (username) {
-                alert('You will now be logged out, please log back in with the updated username');
-                localStorage.clear();
-                window.open("/", "_self")
+            if (updateObject.Username !== user.Username) {
+                alert('Your username has been updated.');
+                localStorage.setItem('user', updateObject.Username);
+                window.open(`/users/${updateObject.Username}`, '_self')
             }
         } catch (error) {
             console.log(error)
@@ -41,16 +35,23 @@ function ProfileView(props) {
     }
 
     const handleDelete = async () => {
-        try {
-            const response = await axios.delete(`https://trackm-app.herokuapp.com/users/${user.Username}`, {
-                headers: { Authorization: `Bearer ${token}` }
-            });
-            alert('Your account was permanently deleted');
-            localStorage.clear();
-            window.open('/', '_self');
-        } catch (error) {
-            console.log(error);
+        if (user.Username && token) {
+            let confirmDelete = confirm('You account will be permanently deleted, are you sure you want to continue?');
+            if (!confirmDelete) return;
+
+            try {
+                await axios.delete(`https://trackm-app.herokuapp.com/users/${user.Username}`, {
+                    headers: { Authorization: `Bearer ${token}` }
+                });
+                alert('Your account was permanently deleted');
+                localStorage.clear();
+                deleteUser({});
+                window.open('/', '_self');
+            } catch (error) {
+                console.log(error);
+            }
         }
+
     }
 
     return (
@@ -66,9 +67,9 @@ function ProfileView(props) {
                     {(!user.FavoriteMovies) ?
                         <div className="main-view"></div>
                         :
-                        <FavoriteMovies favoriteMoviesList={movies.filter((movie) => user.FavoriteMovies.includes(movie._id))} unfavorite={unfavorite} />
+                        <FavoriteMovies favoriteMoviesList={movies.filter((movie) => user.FavoriteMovies.includes(movie._id))} />
                     }
-                    <UpdateUser handleSubmit={handleSubmit} user={user} />
+                    <UpdateUser handleUpdate={handleUpdate} user={user} />
                     <Row>
                         <h2 className="update-title">Delete User Info</h2>
                     </Row>
@@ -90,4 +91,4 @@ const mapStateToProps = state => {
         user: state.user
     };
 }
-export default connect(mapStateToProps, { setUser })(ProfileView);
+export default connect(mapStateToProps, { updateUser, deleteUser })(ProfileView);
